refactor move `GetVcidFromServiceName` from boyar to strelets (single auth for service names) extractg `WatchAndReportServicesStatus` from main/boyar to separate unit (service)
