+++ boyar/helpers.go
@@ -3,9 +3,7 @@ package boyar
 import (
 	"fmt"
 	"github.com/orbs-network/boyarin/strelets"
-	"strconv"
 	"strings"
-	"time"
 )
 
 func buildPeersMap(nodes []*strelets.FederationNode, gossipPort int) *strelets.PeersMap {
@@ -24,24 +22,6 @@ func buildPeersMap(nodes []*strelets.FederationNode, gossipPort int) *strelets.P
 	}
 
 	return &peersMap
-}
-
-func getVcidFromServiceName(serviceName string) int64 {
-	tokens := strings.Split(serviceName, "-")
-	if len(tokens) < 2 {
-		return -1
-	}
-
-	result, err := strconv.ParseInt(tokens[len(tokens)-2], 10, 0)
-	if err != nil {
-		return -1
-	}
-
-	return result
-}
-
-func formatAsISO6801(t time.Time) string {
-	return t.Format(time.RFC3339)
 }
 
 func aggregateErrors(errors []error) error {
