+++ services/report_status.go
@@ -1,15 +1,38 @@
-package boyar
+package services
 
 import (
 	"context"
 	"fmt"
 	"github.com/orbs-network/boyarin/log_types"
+	"github.com/orbs-network/boyarin/strelets"
 	"github.com/orbs-network/boyarin/strelets/adapter"
+	"github.com/orbs-network/boyarin/supervized"
 	"github.com/orbs-network/scribe/log"
 	"time"
 )
 
-func ReportStatus(ctx context.Context, logger log.Logger, since time.Duration) error {
+const SERVICE_STATUS_REPORT_PERIOD = 1 * time.Minute
+const SERVICE_STATUS_REPORT_TIMEOUT = 30 * time.Second
+
+func WatchAndReportServicesStatus(logger log.Logger) chan interface{} {
+	return supervized.GoForever(func() {
+		for {
+			start := time.Now()
+			ctx, cancel := context.WithTimeout(context.Background(), SERVICE_STATUS_REPORT_TIMEOUT)
+			if err := reportStatus(ctx, logger, SERVICE_STATUS_REPORT_PERIOD); err != nil {
+				logger.Error("status check failed", log.Error(err))
+			}
+			cancel()
+			<-time.After(SERVICE_STATUS_REPORT_PERIOD - time.Since(start)) // to report exactly every minute
+		}
+	})
+}
+
+func formatAsISO6801(t time.Time) string {
+	return t.Format(time.RFC3339)
+}
+
+func reportStatus(ctx context.Context, logger log.Logger, since time.Duration) error {
 	// We really don't need any options here since we're just observing
 	orchestrator, err := adapter.NewDockerSwarm(adapter.OrchestratorOptions{})
 	if err != nil {
@@ -30,7 +53,7 @@ func ReportStatus(ctx context.Context, logger log.Logger, since time.Duration) e
 			log.String("createdAt", formatAsISO6801(s.CreatedAt)),
 		}
 
-		if vcid := getVcidFromServiceName(s.Name); vcid > 0 {
+		if vcid := strelets.GetVcidFromServiceName(s.Name); vcid > 0 {
 			fields = append(fields, log_types.VirtualChainId(vcid))
 		}
